% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate_vec}
\alias{translate_vec}
\title{Translate elements in a character vector}
\usage{
translate_vec(
  words,
  rosetta_stone,
  from = "french",
  to = "english",
  category_selection = "."
)
}
\arguments{
\item{words}{A character vector containing the words to be translated.}

\item{rosetta_stone}{A tibble that serves as a translation dictionary, with columns
named after the languages it supports and a 'category' column.}

\item{from}{The name of the column in `rosetta_stone` representing the source language.}

\item{to}{The name of the column in `rosetta_stone` representing the target language.}

\item{category_selection}{Optional category selection to filter rows in `rosetta_stone`.}
}
\value{
A character vector with translated elements.
}
\description{
This function takes a character vector, a dataset that acts as a translation
dictionary, and optional arguments specifying the languages to translate from
and to. It returns a modified character vector with elements translated where possible.
}
\details{
The function filters `rosetta_stone` based on `category_selection`, then uses it to
create a named vector. Elements in `words` are replaced by their corresponding values
in the named vector, if they exist.
}
\examples{
\dontrun{
rosetta <- tibble(french = c("le", "une"), english = c("the", "a"), category = "articles")
translate_vec(c("le", "a", "une", "other"),
rosetta,
from = "french",
to = "english",
category_selection = "articles")
}

}
