% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{Translate}
\alias{Translate}
\title{DNA or RNA translation into amino acid code}
\usage{
Translate(x)
}
\arguments{
\item{x}{a text string containig a DNA or RNA sequence}
}
\value{
A a text tring that returns a Translated seq
}
\description{
DNA or RNA translation into amino acid code
}
\examples{
x <- "ATGTCCTTAGGCT"
Translate(x)
}
